// backend/src/common/types.ts
export enum Role {
  CUSTOMER = 'CUSTOMER',
  TAILOR = 'TAILOR',
  DESIGNER = 'DESIGNER',
  ADMIN = 'ADMIN'
}

export enum Category {
  SHIRT = 'SHIRT',
  PANT = 'PANT',
  BLAZER = 'BLAZER'
}

export enum OrderStatus {
  PAID = 'PAID',
  ASSIGNED = 'ASSIGNED',
  PRODUCTION = 'PRODUCTION',
  QC = 'QC',
  SHIPPED = 'SHIPPED',
  DELIVERED = 'DELIVERED'
}